requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asttokens=2.0.5=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.10.0=pypi_0
black=21.12b0=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.6.15=ha878542_0
cairo=1.16.0=ha61ee94_1012
certifi=2021.10.8=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.0.3=pypi_0
cloudpickle=2.0.0=pypi_0
cudatoolkit=11.3.1=h9edb442_10
cycler=0.11.0=pypi_0
dask=2021.11.2=pypi_0
debugpy=1.5.1=pypi_0
decorator=5.1.1=pypi_0
dominate=2.6.0=pypi_0
efficientnet-pytorch=0.6.3=pypi_0
entrypoints=0.3=pypi_0
enum-compat=0.0.3=pypi_0
executing=0.8.2=pypi_0
expat=2.4.8=h27087fc_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=h8e229c2_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.28.5=pypi_0
freetype=2.12.1=hca18f0e_0
fsspec=2022.2.0=pypi_0
future=0.18.2=pypi_0
gdk-pixbuf=2.42.8=hff1cb4f_0
gettext=0.19.8.1=h73d1719_1008
glib=2.72.1=h6239696_0
glib-tools=2.72.1=h6239696_0
h5py=3.6.0=pypi_0
icu=70.1=h27087fc_0
idna=3.3=pypi_0
imageio=2.14.0=pypi_0
ipykernel=6.7.0=pypi_0
ipython=8.0.1=pypi_0
jedi=0.18.1=pypi_0
joblib=1.1.0=pypi_0
jpeg=9e=h166bdaf_2
jsonpatch=1.32=pypi_0
jsonpointer=2.2=pypi_0
jupyter-client=7.1.1=pypi_0
jupyter-core=4.9.1=pypi_0
kiwisolver=1.3.2=pypi_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=4.0.0=h27087fc_0
libdeflate=1.13=h166bdaf_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libglib=2.72.1=h2d90d5f_0
libgomp=12.1.0=h8d9b700_16
libiconv=1.16=h516909a_0
libnsl=2.0.0=h7f98852_0
libpng=1.6.37=h753d276_4
libsqlite=3.39.2=h753d276_1
libstdcxx-ng=12.1.0=ha89aaad_16
libtiff=4.4.0=h0e0dad5_3
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.4=h166bdaf_0
libxcb=1.13=h7f98852_1004
libxml2=2.9.14=h22db469_4
libzlib=1.2.12=h166bdaf_2
llvmlite=0.36.0=pypi_0
locket=0.2.1=pypi_0
lxml=4.7.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
munch=2.5.0=pypi_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.4=pypi_0
networkx=2.6.3=pypi_0
numba=0.53.1=pypi_0
numpy=1.22.0=pypi_0
opencv-python=4.5.3.56=pypi_0
openjpeg=2.5.0=h7d73246_1
openslide=3.4.1=h71beb9a_5
openslide-python=1.2.0=pypi_0
openssl=3.0.5=h166bdaf_1
packaging=21.3=pypi_0
pandas=1.4.0=pypi_0
parso=0.8.3=pypi_0
partd=1.2.0=pypi_0
pathspec=0.9.0=pypi_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=pypi_0
pip=22.2.2=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
platformdirs=2.4.1=pypi_0
pretrainedmodels=0.7.4=pypi_0
prompt-toolkit=3.0.24=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.1=pypi_0
pyaml=21.10.1=pypi_0
pygments=2.11.2=pypi_0
pyparsing=3.0.7=pypi_0
python=3.9.13=h2660328_0_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.9=2_cp39
pytz=2021.3=pypi_0
pywavelets=1.2.0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=pypi_0
readline=8.1.2=h0f457ee_0
requests=2.26.0=pypi_0
scikit-image=0.18.3=pypi_0
scikit-learn=1.0.2=pypi_0
scikit-optimize=0.9.0=pypi_0
scipy=1.7.3=pypi_0
segmentation-models-pytorch=0.2.1=pypi_0
setuptools=65.2.0=py39hf3d152e_0
six=1.16.0=pypi_0
soupsieve=2.3.1=pypi_0
sqlite=3.39.2=h4ff8645_1
stack-data=0.1.4=pypi_0
tensorboardx=2.4.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
timm=0.4.12=pypi_0
tk=8.6.12=h27826a3_0
tomli=1.2.3=pypi_0
toolz=0.11.2=pypi_0
torch=1.10.2+cu113=pypi_0
torch-model-archiver=0.5.2=pypi_0
torch-summary=1.4.5=pypi_0
torch-workflow-archiver=0.2.2=pypi_0
torchaudio=0.10.2+cu113=pypi_0
torchfile=0.1.0=pypi_0
torchserve=0.5.2=pypi_0
torchvision=0.11.3+cu113=pypi_0
tornado=6.1=pypi_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pypi_0
tzdata=2022c=h191b570_0
urllib3=1.26.8=pypi_0
visdom=0.1.8.9=pypi_0
wcwidth=0.2.5=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h166bdaf_0
zlib=1.2.12=h166bdaf_2
zstd=1.5.2=h6239696_4
